chore: move violation type to seprate file
