+++ src/validator/constraints/other/compound.ts
@@ -1,4 +1,5 @@
-import type { ConstraintContext, FieldRules, Violation } from '../../types';
+import type { ConstraintContext, FieldRules } from '../../types';
+import { Violation } from '@/validator/violation';
 
 export function compound(rules: FieldRules) {
   return function compoundConstraint(value: unknown, context: ConstraintContext): Violation[] {
