+++ src/validator/factory/create-validator.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest';
 import { UnknownConstraintError } from '../errors';
-import type { ConstraintContext, Violation } from '../types';
+import type { ConstraintContext } from '../types';
 import { createValidator } from './create-validator';
+import { Violation } from '@/validator/violation';
 
 describe('Validator', () => {
   describe('Registry', () => {
