+++ src/validator/index.ts
@@ -3,3 +3,4 @@ export * from './types';
 export * from './constraints';
 export * from './flatten-violations';
 export { createValidationMiddleware } from './middleware/create-validation-middleware';
+export type { Violation } from '@/validator/violation';
