+++ src/validator/violation.ts
@@ -1,0 +1,6 @@
+export type Violation = {
+  path: string;
+  message: string;
+  constraint: string;
+  value: unknown;
+};
